requirements.txt
pyautogui
pymupdf
psycopg2-binary
pandas
python-dotenv